backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
trimesh>=4.0.0
shap>=0.44.0
psycopg2-binary>=2.9.0
openai>=1.12.0
joblib>=1.3.0
pydantic>=2.5.0
bcrypt>=4.1.0
PyJWT>=2.8.0
email-validator>=2.1.0
